refactor: 포맷 유틸 함수를 utils/format.ts로 분리
- formatDate: 날짜 포맷팅 (YYYY-MM-DD → YYYY.MM.DD)
- formatSeatInfo: 좌석 정보 포맷팅 (층/구역/열/번)
- formatActorNames: 배우 이름 목록 포맷팅
