+++ src/app/(auth)/review-detail/[id].tsx
@@ -15,6 +15,7 @@ import { ReviewInfoSection } from '@/components/ReviewInfoSection'
 import { TicketCard } from '@/components/TicketCard'
 import { toast } from '@/components/Toaster'
 import { useUser } from '@/providers/user.provider'
+import { formatActorNames, formatDate, formatSeatInfo } from '@/utils/format'
 import { showPointRewardToast } from '@/utils/pointReward'
 import { useQuery } from '@tanstack/react-query'
 import { useLocalSearchParams, useRouter } from 'expo-router'
@@ -34,33 +35,6 @@ const FACILITY_LEVEL_LABELS: Record<number, string> = {
   1: '좋지 않아요',
   2: '보통이에요',
   3: '쾌적해요',
-}
-
-// 좌석 정보 포맷팅
-function formatSeatInfo(ticket: {
-  floor?: number
-  zone?: string
-  col?: string
-  number?: string
-}): string {
-  const parts: string[] = []
-  if (ticket.floor) parts.push(`${ticket.floor}층`)
-  if (ticket.zone) parts.push(`${ticket.zone}구역`)
-  if (ticket.col) parts.push(`${ticket.col}열`)
-  if (ticket.number) parts.push(`${ticket.number}번`)
-  return parts.join(' ') || '좌석 정보 없음'
-}
-
-// 배우 이름 포맷팅
-function formatActorNames(actors?: { name?: string }[]): string {
-  if (!actors || actors.length === 0) return ''
-  return actors.map((a) => a.name).filter(Boolean).join(' ')
-}
-
-// 날짜 포맷팅 (YYYY-MM-DD → YYYY.MM.DD)
-function formatDate(dateStr?: string): string {
-  if (!dateStr) return ''
-  return dateStr.replace(/-/g, '.')
 }
 
 export default function ReviewDetail() {
